remove debug draw and fix Rendering transformations

+++ Source/ModuleRenderTransformation.cpp
@@ -38,9 +38,10 @@ bool ModuleRenderTransformation::Init()
 
 	// Define the model, view and projection matrix
 	float4x4 proj = frustum.ProjectionMatrix();
-	float4x4 model = float4x4::FromTRS(float3(0.0f, 0.0f, -4.0f), float3x3::RotateY((float)pi / 4.0f), float3(1.0f, 1.0f, 1.0f));
+	float4x4 model = float4x4::FromTRS(float3(0.0f, 0.0f, -4.0f), float3x3::RotateY(pi / 4.f), float3(1.0f, 1.0f, 1.0f));
 	//TODO: Use custom LookAt function ->  LookAt(float3(0.0f, 4.0f, 8.0f), float3(0.0f, 0.0f, 0.0f), float3::unitY);
 	float4x4 view = frustum.ViewMatrix();
+	//float4x4 view = float4x4::LookAt(float3(0.f, 0.0f, -1.0f), float3(0.0f, 0.0f, -1.0f), float3(0.0f, 1.0f, 0.0f), float3(0.0f, 1.0f, 0.0f));
 
 	// Bind the VBO
 	glUniformMatrix4fv(0, 1, GL_TRUE, &model[0][0]);
